refactor(user-account): modularize UserAccount component with new tabs for personal info, consultant, and lab appointments
- Refactored the UserAccount component to utilize separate tabs for Personal Info, Consultant Appointments, Lab Appointments, and Settings, enhancing organization and user experience.
- Introduced new components for each tab, improving code maintainability and readability.
- Updated CSS styles for the new tab layout and individual components, ensuring a cohesive design.
- Removed deprecated styles and imports, streamlining the CSS file.
- Enhanced appointment management with improved UI for displaying and interacting with appointment details.
