+++ src/pages/UserAccount/components/PersonalInfoTab/PersonalInfoTab.css
@@ -1,0 +1,98 @@
+/* Profile Header */
+.profile-header {
+  background: linear-gradient(135deg, #667eea 0%, #764ba2 100%);
+  border: none !important;
+  border-radius: 20px !important;
+  margin-bottom: 24px;
+  box-shadow: 0 8px 32px rgba(102, 126, 234, 0.3) !important;
+  color: white;
+}
+
+.profile-header .ant-card-body {
+  padding: 32px;
+}
+
+.avatar-section {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.user-avatar {
+  background: linear-gradient(45deg, #ff6b6b, #ee5a24) !important;
+  border: 4px solid rgba(255, 255, 255, 0.3);
+  box-shadow: 0 4px 20px rgba(0, 0, 0, 0.2);
+  cursor: pointer;
+}
+
+.avatar-edit-btn {
+  background: #667eea !important;
+  border: 2px solid white !important;
+  color: white !important;
+}
+
+.avatar-edit-btn:hover {
+  background: #5a67d8 !important;
+  transform: scale(1.1);
+}
+
+.profile-info {
+  text-align: left;
+}
+
+.user-name {
+  color: white !important;
+  margin-bottom: 8px !important;
+  font-weight: 700 !important;
+}
+
+.user-email {
+  color: rgba(255, 255, 255, 0.9) !important;
+  font-size: 16px !important;
+  display: block;
+  margin-bottom: 16px;
+}
+
+.profile-completion {
+  display: flex;
+  align-items: center;
+  color: rgba(255, 255, 255, 0.9);
+}
+
+.profile-completion .ant-progress-bg {
+  background: linear-gradient(to right, #ff6b6b, #feca57) !important;
+}
+
+.profile-stats {
+  text-align: center;
+}
+
+.profile-stats .ant-statistic-title {
+  color: rgba(255, 255, 255, 0.8) !important;
+  font-size: 14px !important;
+}
+
+.profile-stats .ant-statistic-content {
+  color: white !important;
+  font-size: 20px !important;
+  font-weight: 600 !important;
+}
+
+.edit-profile-btn {
+  margin-top: 16px;
+  width: 100%;
+  height: 45px;
+  border-radius: 25px !important;
+  background: rgba(255, 255, 255, 0.2) !important;
+  border: 2px solid rgba(255, 255, 255, 0.3) !important;
+  color: white !important;
+  font-weight: 600 !important;
+  backdrop-filter: blur(10px);
+  transition: all 0.3s ease !important;
+}
+
+.edit-profile-btn:hover {
+  background: rgba(255, 255, 255, 0.3) !important;
+  border-color: rgba(255, 255, 255, 0.5) !important;
+  transform: translateY(-2px);
+}
